Added integration testing for all SqlDAO files

+++ final-capstone/java/src/main/java/com/techelevator/dao/BrewerySqlDAO.java
@@ -14,7 +14,7 @@ import java.util.List;
 
 @Service
 @Component
-public class BrewerySqlDAO implements BreweryDAO{
+public class BrewerySqlDAO implements BreweryDAO {
 
     private JdbcTemplate jdbcTemplate;
 
@@ -90,7 +90,7 @@ public class BrewerySqlDAO implements BreweryDAO{
 
         brewery.setId(results.getLong("brewery_id"));
         brewery.setBreweryName(results.getString("brewery_name"));
-        brewery.setBrewerId(results.getInt("brewer_id"));
+        brewery.setBrewerId(results.getLong("brewer_id"));
         brewery.setBreweryStreetAddress(results.getString("brewery_street_address"));
         brewery.setBreweryCity(results.getString("brewery_city"));
         brewery.setBreweryState(results.getString("brewery_state"));
